requirements.txt
setuptools==80.9.0
wheel==0.45.1
python-dotenv==1.1.1